+++ spec/unit/generator/xcconfig/pod_xcconfig_spec.rb
@@ -129,7 +129,7 @@ module Pod
           it 'sets the PODS_TARGET_SRCROOT build variable for local pod' do
             @pod_target.sandbox.store_local_path(@pod_target.pod_name, @spec.defined_in_file)
             @xcconfig = @generator.generate
-            @xcconfig.to_hash['PODS_TARGET_SRCROOT'].should == '${PODS_ROOT}/../../spec/fixtures/banana-lib/BananaLib.podspec'
+            @xcconfig.to_hash['PODS_TARGET_SRCROOT'].should == '${PODS_ROOT}/../../spec/fixtures/banana-lib'
           end
 
           it 'adds the library build headers and public headers search paths to the xcconfig, with quotes' do